fixed order

+++ services/order/src/main/java/com/pc/order/dto/PurchaseResponse.java
@@ -3,10 +3,10 @@ package com.pc.order.dto;
 import java.math.BigDecimal;
 
 public record PurchaseResponse(
-        Integer productId,
+        Long productId,
         String name,
         String description,
         BigDecimal price,
-        double quantity
+        Integer quantity
 ) {
 }
