+++ services/order/src/main/java/com/pc/order/repository/OrderLineRepository.java
@@ -4,7 +4,7 @@ import com.pc.order.model.OrderLine;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderLineRepository extends JpaRepository<OrderLine, Integer> {
+public interface OrderLineRepository extends JpaRepository<OrderLine, Long> {
 
-    List<OrderLine> findAllByOrderId(Integer orderId);
+    List<OrderLine> findAllByOrderId(Long orderId);
 }
